Added changes and fixes from pull request review.
