+++ controllers/login.js
@@ -1,5 +1,5 @@
-const logIn = require("../public/javascripts/data/login");
-const encryption = require("../public/javascripts/encryption");
+const logIn = require("../public/javascripts/data/user");
+// const hashing = require("../public/javascripts/hashing");
 
 
 // GET request
@@ -14,13 +14,14 @@ exports.checkLogin = function(req, res, next) {
   let username = req.body.username;
   let password = req.body.password;
 
-  if(logIn.tryLogin(username, password)) {
-    // TODO:
-    // 1. figure out how to use express-session
-    // 2. assign some type of global object to this user?
-    res.redirect('/');
+  // If username and/or password was incorrect
+  if(!logIn.tryLogin(username, password)) {
+
+    res.render('login', { errorMessage: "Wrong username and/or password. Try again" })
   }
 
-  // If username and/or password was incorrect
-  res.render('login', { errorMessage: "Wrong username and/or password. Try again" })
+  // TODO:
+  // 1. figure out how to use express-session
+  // 2. assign some type of global object to this user?
+  res.redirect('/');
 };
