+++ routes/index.js
@@ -5,11 +5,11 @@ let index = require('../controllers/index');
 let logIn = require('../controllers/login');
 let signUp = require('../controllers/signup')
 
-/* home page. */
+// home page
 router.get('/', index.index);
 router.post('/', index.add_paragraph);
 
-/* login page. */
+// login page
 router.get('/login', logIn.log_in);
 router.post('/login', logIn.checkLogin)
