+++ views/login.pug
@@ -1,7 +1,7 @@
 extends layout 
 block content 
     h1 Hey there guy. Login
-    form(action='http://127.0.0.1:3000/login' method='POST')
+    form(action='/login' method='POST')
         // user input
         | Username:
         br
@@ -18,5 +18,5 @@ block content
     br
     p= errorMessage
     br
-    p Not a member yet bubby? 
-    a(href="http://127.0.0.1:3000/signup") Sign up here
+    p Not a member yet buddy? 
+    a(href="/signup") Sign up here
