+++ views/signup.pug
@@ -1,7 +1,7 @@
 extends layout 
 block content 
     h1 Hey there buddy, Sign up
-    form(action='http://127.0.0.1:3000/signup' method='POST')
+    form(action='/signup' method='POST')
         // user input
         | Preferred Username:
         br
@@ -17,3 +17,6 @@ block content
         input(type='submit' value='signup')
         br
         p= errorMessage
+        br
+        p Already a member, guy? 
+        a(href="/login") Sign in here
